pydrake: Rename `_math` to `math`
