+++ bindings/pydrake/BUILD.bazel
@@ -83,7 +83,7 @@ drake_py_library(
 
 drake_pybind_library(
     name = "math_py",
-    cc_so_name = "_math",
+    cc_so_name = "math",
     cc_srcs = ["math_py.cc"],
     package_info = PACKAGE_INFO,
     py_deps = [
@@ -222,19 +222,11 @@ drake_py_test(
 )
 
 drake_py_test(
-    # Note the non-standard name due to #8041.
-    name = "test/math_test",
+    name = "math_test",
     size = "small",
-    srcs = ["test/math_test.py"],
-    main = "test/math_test.py",
     deps = [
         ":math_py",
     ],
-)
-
-alias(
-    name = "math_test",
-    actual = "test/math_test",
 )
 
 drake_py_test(
