+++ bindings/pydrake/systems/controllers_py.cc
@@ -37,7 +37,7 @@ PYBIND11_MODULE(controllers, m) {
   // NOLINTNEXTLINE(build/namespaces): Emulate placement in namespace.
   using namespace drake::systems::controllers;
 
-  py::module::import("pydrake._math");
+  py::module::import("pydrake.math");
   py::module::import("pydrake.symbolic");
   py::module::import("pydrake.systems.primitives");
